+++ forms/UgovorFilesForm.cs
@@ -33,7 +33,7 @@ namespace CSharp_SQL_App.forms {
             OleDbCommand command;
             connection = GetConnection();
             connection.Open();
-            String query = "SELECT * FROM ugovorFiles WHERE uGuid = @uGuid ORDER BY id";      
+            String query = "SELECT * FROM ugovorFiles WHERE uGuid = @uGuid ORDER BY id";
             command = new OleDbCommand(query, connection);
             command.Parameters.AddWithValue("@uGuid", uGuid);
             dt = new DataTable();
@@ -51,10 +51,10 @@ namespace CSharp_SQL_App.forms {
             if (dataGridViewUgovorFiles.SelectedRows.Count > 0) {
                 filePath = dataGridViewUgovorFiles.SelectedRows[0].Cells["fajlPutanja"].Value.ToString();
                 System.Diagnostics.Process.Start(filePath);
-            }         
+            }
         }
 
-        private void buttonDodaj_Click(object sender, EventArgs e) {          
+        private void buttonDodaj_Click(object sender, EventArgs e) {
             OpenFileDialog openFileDialog = new OpenFileDialog();
             openFileDialog.Multiselect = false;
             openFileDialog.ShowDialog();
@@ -73,22 +73,25 @@ namespace CSharp_SQL_App.forms {
             }
             fillUgovorFilesDataGrid();
         }
-
-        private void buttonObrisi_Click(object sender, EventArgs e) {
-            if (dataGridViewUgovorFiles.SelectedRows.Count > 0) {
-                String selectedFilePath = dataGridViewUgovorFiles.SelectedRows[0].Cells["fajlPutanja"].Value.ToString();
-                OleDbConnection connection;
-                OleDbCommand command;
-                connection = GetConnection();
-                connection.Open();
-                String query = "DELETE FROM ugovorFiles WHERE fajlPutanja = @fajlPutanja";
-                command = new OleDbCommand(query, connection);
-                command.Parameters.AddWithValue("@fajlPutanja", selectedFilePath);
-                command.ExecuteNonQuery();
-                connection.Close();
-                fillUgovorFilesDataGrid();
-            }    
+        private void buttonUkloni_Click(object sender, EventArgs e) {
+            ConfirmationForm confirmationForm = new ConfirmationForm();
+            if (confirmationForm.ShowDialog().Equals(DialogResult.Yes)) {
+                if (dataGridViewUgovorFiles.SelectedRows.Count > 0) {
+                    String selectedFilePath = dataGridViewUgovorFiles.SelectedRows[0].Cells["fajlPutanja"].Value.ToString();
+                    OleDbConnection connection;
+                    OleDbCommand command;
+                    connection = GetConnection();
+                    connection.Open();
+                    String query = "DELETE FROM ugovorFiles WHERE fajlPutanja = @fajlPutanja";
+                    command = new OleDbCommand(query, connection);
+                    command.Parameters.AddWithValue("@fajlPutanja", selectedFilePath);
+                    command.ExecuteNonQuery();
+                    connection.Close();
+                    fillUgovorFilesDataGrid();
+                }
+            }
         }
+
         private void buttonZatvori_Click(object sender, EventArgs e) {
             this.Close();
         }
